+++ app/code/local/Radarsofthouse/Reepay/Model/System/Config/Source/Dropdown/Allowwedpayment.php
@@ -37,6 +37,7 @@ class Radarsofthouse_Reepay_Model_System_Config_Source_Dropdown_Allowwedpayment
             array('value' => 'paypal', 'label' => __('PayPal')),
             array('value' => 'resurs', 'label' => __('Resurs Bank')),
             array('value' => 'vipps', 'label' => __('Vipps')),
+            array('value' => 'googlepay', 'label' => __('Google Pay'))
         );
     }
 }